feat(server): properly returning 422 HTTP code for syntax errors
Right now we're returning a generic 500 error for any error related to
tentaclesql. We're interested on discarding all the user facing errors
from this service consumers. So this PR fixes that by analysing the
error returned by executor and returning a unprocessable_entity error if
the privded SQL has a syntaxis error.