fix: make SqliteError fail nicely (#22)
Our users can make mistakes. For instance, they may use a custom field
that is deleted, or mistype `employe`. Those are invalid queries, hence
it should return a 400 error and give back the information to them.

With the current implementation, we were only returning a 400 when we
failed to parse the SQL. But valid SQLs can be invalid queries.

This will mitigate hundreds of errors everyday and improve the user
experience.